Split shema into files
