+++ src/server/db/index.ts
@@ -2,7 +2,8 @@ import { neon, neonConfig } from '@neondatabase/serverless';
 import { drizzle } from 'drizzle-orm/neon-http';
 
 import { env } from "@/env.mjs";
-import * as schema from "./schema";
+import * as data from "./schema/data";
+import * as users from "./schema/users";
 
 
 neonConfig.fetchConnectionCache = true;
@@ -10,5 +11,8 @@ neonConfig.fetchConnectionCache = true;
 const sql = neon(env.DATABASE_URL);
 
 export const db = drizzle(sql, {
-  schema
+  schema: {
+    ...data,
+    ...users,
+  }
 });
