+++ src/server/db/schema/table.ts
@@ -1,0 +1,9 @@
+import { pgTableCreator } from "drizzle-orm/pg-core";
+
+/**
+ * This is an example of how to use the multi-project schema feature of Drizzle ORM. Use the same
+ * database instance for multiple projects.
+ *
+ * @see https://orm.drizzle.team/docs/goodies#multi-project-schema
+ */
+export const myPgTable = pgTableCreator((name) => `shine_${name}`);
